requirements.txt
cachetools==3.1.1
certifi==2019.6.16
chardet==3.0.4
Django==2.2.3
django-cors-headers==3.1.0
djangorestframework==3.10.1
google-api-core==1.14.0
google-auth==1.6.3
google-cloud-core==1.0.2
google-cloud-translate==1.6.0
googleapis-common-protos==1.6.0
grpcio==1.22.0
idna==2.8
protobuf==3.9.0
psycopg2==2.8.3
pyasn1==0.4.5
pyasn1-modules==0.2.5
pytz==2019.1
requests==2.22.0
rsa==4.0
selenium==3.141.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
tensorflow
indic_transliteration